feat: add so much features

+++ src/main/java/ru/nskopt/services/CategoryService.java
@@ -1,11 +1,11 @@
 package ru.nskopt.services;
 
 import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.nskopt.models.Category;
+import ru.nskopt.exceptions.ResourceNotFoundException;
+import ru.nskopt.models.entities.Category;
 import ru.nskopt.repositories.CategoryRepository;
 
 @Slf4j
@@ -19,16 +19,22 @@ public class CategoryService {
     return categoryRepository.findAll();
   }
 
-  public Optional<Category> findById(Long id) {
-    return categoryRepository.findById(id);
+  public Category findById(Long id) {
+    return categoryRepository
+        .findById(id)
+        .orElseThrow(() -> new ResourceNotFoundException("Category with id " + id + " not found"));
   }
 
   public Category save(Category category) {
     log.info("Save {}", category);
+
     return categoryRepository.save(category);
   }
 
   public void deleteById(Long id) {
+    if (!categoryRepository.existsById(id))
+      throw new ResourceNotFoundException("Category with id " + id + " not found");
+
     log.info("Delete category with id {}", id);
     categoryRepository.deleteById(id);
   }
